requirements.txt
# Agent Core Runtime 의존성
# Docker 컨테이너로 배포됨 (Lambda 아님)

# AWS SDK
boto3>=1.35.0
botocore>=1.35.0

# Strands Agents
strands-agents>=0.1.0
strands-agents-tools

# Core dependencies
pydantic>=2.0.0
pyyaml

# HTTP Client (API 호출용 - weekly_report)
httpx>=0.26.0

# Environment
python-dotenv>=1.0.0

# Agent Core (배포용)
bedrock-agentcore-starter-toolkit

# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.1.0